app/requirements.txt
blinker==1.4
flask==0.12.2
gevent==1.2.2
itsdangerous==0.24
mongoengine==0.15.0
mongomock==3.8.0
flake8==3.5.0
pylint==1.8.1
pytest-flask==0.10.0
pytest-xdist==1.20.1
pytest==3.3.1
raven[flask]==6.4.0
requests==2.18.4
teamcity-messages==1.21
voluptuous==0.10.5
uwsgi==2.0.15
boto==2.48.0
requests_mock==1.4.0
python-telegram-bot==9.0.0
flask-cors
pysftp==0.2.9
pyonfido